+++ client/test/api_internal_test.go
@@ -94,6 +94,18 @@ func Test_taikuncore_InternalAPIService(t *testing.T) {
 
 	})
 
+	t.Run("Test InternalAPIService InternalProjectsForBilling", func(t *testing.T) {
+
+		t.Skip("skip test")  // remove to run test
+
+		resp, httpRes, err := apiClient.InternalAPI.InternalProjectsForBilling(context.Background()).Execute()
+
+		require.Nil(t, err)
+		require.NotNil(t, resp)
+		assert.Equal(t, 200, httpRes.StatusCode)
+
+	})
+
 	t.Run("Test InternalAPIService InternalProjectsList", func(t *testing.T) {
 
 		t.Skip("skip test")  // remove to run test